requirements.txt
certifi==2020.6.20              # For secure HTTP requests
Django==5.1.7                   # Core framework for REST API
django-cors-headers==4.7.0      # For handling CORS in API
django-environ==0.12.0          # For environment variable management
django-filter==24.3             # Filtering support for API
django-leaflet==0.31.0          # GIS mapping with Leaflet
djangorestframework==3.15.2     # REST API framework
djangorestframework-gis==1.1    # GIS support for REST framework
djangorestframework-simplejwt==5.3.1  # JWT authentication for DRF
GDAL==3.4.1                     # Geospatial data abstraction library
gunicorn==21.2.0                # WSGI server for deployment
numpy==1.26.4                  # Machine learning and numerical operations
pillow==11.1.0                  # Image processing (if needed for GIS/ML)
psycopg2-binary==2.9.10         # PostgreSQL adapter (common with GIS)
PyJWT==2.3.0                    # JSON Web Token implementation
requests==2.25.1                # For making HTTP requests
sqlparse==0.4.2                 # Dependency for Django
typing_extensions==4.12.2       # Type hints support
urllib3==1.26.5                 # HTTP client for requests
cryptography==3.4.8             # For secure token handling and encryptiongoogle-auth==2.38.0
google-auth==2.38.0
django-allauth==65.7.0
django-rest-auth==0.9.5
boto3==1.38.13
django-storages==1.14.6
drf_spectacular==0.28.0
